Allow codeblocks in answer choices
This allows for code options in multiple choice using backticks, either code fences
or single backticks. Both for now go to the `<code>` tag, but theoretically
if pl ever supports it, triple backticks could be converted to `<pl-code>` tags

This uses a regex substitution to force parity on opening and closing. it has been checked briefly, but more rigorous testing of it would not hurt

Closes #35